updated: refactored code and removed unnecessary code.  updated readme.
