--- lib/vaahextendflutter/environment/controllers/env_controller.dart
@@ -1,24 +1,0 @@
-import 'package:get/get.dart';
-
-import '../../../env.dart';
-import '../env_helpers.dart';
-
-class EnvController extends GetxController {
-  EnvironmentConfig _config = devEnvConfig;
-  EnvironmentConfig get config => _config;
-
-  EnvController(String environment) {
-    switch (environment) {
-      case 'prod':
-        _config = prodEnvConfig;
-        break;
-      case 'stag':
-        _config = stagEnvConfig;
-        break;
-      default:
-        _config = devEnvConfig;
-        break;
-    }
-    update();
-  }
-}
