+++ lib/vaahextendflutter/environment/env_helpers.dart
@@ -4,9 +4,9 @@
 
 String getEnvFromCommandLine() {
   String environment =
-      const String.fromEnvironment('environment', defaultValue: 'dev');
-  if (!(environment == 'prod' || environment == 'stag')) {
-    environment = 'dev';
+      const String.fromEnvironment('environment', defaultValue: 'development');
+  if (!(environment == 'production' || environment == 'staging')) {
+    environment = 'development';
   }
   return environment;
 }
@@ -15,16 +15,10 @@ String getEnvFromCommandLine() {
 // Main.dart helpers ends
 // ****************************************
 
-class AppVersion {
-  final String build; // build no format '2022010101'
-  final String version; // version format 1.0.0 (major.minor.patch)
-
-  const AppVersion({required this.build, required this.version});
-}
-
 class EnvironmentConfig {
   final String envType;
-  final AppVersion version;
+  final String version;
+  final String build;
   final String baseUrl;
   final String apiBaseUrl;
   final String analyticsID;
@@ -34,6 +28,7 @@ class EnvironmentConfig {
   const EnvironmentConfig({
     required this.envType,
     required this.version,
+    required this.build,
     required this.baseUrl,
     required this.apiBaseUrl,
     required this.analyticsID,
